feat: StartLinkAsDepot + refactor package name
